Split files + update architecture
