+++ constants.py
@@ -1,0 +1,17 @@
+# Constantes pour le jeu Morpion
+
+# Taille de l'écran de jeu
+SCREEN_SIZE = 300
+
+# Taille d'une case de la grille
+SQUARE_SIZE = SCREEN_SIZE // 3
+
+# Épaisseur des lignes de la grille
+LINE_WIDTH = 10
+
+# Couleurs utilisées dans le jeu
+LINE_COLOR = (0, 0, 0)       # Couleur des lignes de la grille
+X_COLOR = (66, 66, 66)       # Couleur du symbole X
+O_COLOR = (28, 170, 156)     # Couleur du symbole O
+VICTORY_COLOR = (255, 0, 0)  # Couleur de la ligne de victoire
+WHITE = (255, 255, 255)      # Couleur de fond
