+++ main.py
@@ -1,0 +1,4 @@
+import game
+
+if __name__ == "__main__":
+    game.jeu()
